requirements.txt
# Core libraries
fastapi>=0.103.1
uvicorn>=0.23.2
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.3.0

# LangChain and Google AI
langchain>=0.0.267
langchain-community>=0.0.16
langchain-google-genai>=0.0.5
google-generativeai>=0.3.1

# Vector store and embedding
faiss-cpu>=1.7.4
pypdf>=3.15.1

# Telegram bot
pyTelegramBotAPI>=4.14.0

# Gradio for the UI
gradio>=4.0.0

# Additional utilities
requests>=2.31.0
tqdm>=4.66.1